+++ src/main/java/com/lothrazar/cyclic/util/EntityUtil.java
@@ -100,7 +100,7 @@ public class EntityUtil {
   }
 
   private static void teleportWallSafe(LivingEntity player, Level world, double x, double y, double z) {
-    BlockPos coords = new BlockPos(x, y, z);
+    BlockPos coords = new BlockPos((int) x, (int) y, (int) z);
     world.getChunk(coords).setUnsaved(true);
     player.teleportTo(x, y, z);
     moveEntityWallSafe(player, world);
